Bootstrap typescript-webcomponent based application
Basic setup of all the boilerplate needed for developing web components
and a tiny html page to host them in a dev application.

Typescript for type safety. Eslint and prettier for standardisation.

lit-element to simplify component creation, mostly for templating safety
and backwards compatibility if needed.

Using web-test-runner/web-dev-server for fast development, they are
arbitrary choices that can easily be swapped out.

Testing is done with @open-wc/testing which is really just a convince
wrapper around chai with some niceties to help with testing web
components.