Render ticker rates from blockchain.info API
This basically just uses the rate component to render the entire
response from the blockchain api. I added some state types to make sure
we are always represent which state we are in and capture all
possibilities.

The API is very basic so no complex logic needs to be done outside the
component.