Add update methods to ticker
From the spec: "This information changes every few seconds" so the
ticker will refetch from the api automatically every 30 seconds. Also
provided is an update button in case the refresh time is too long for a
user.

The previous data is still displayed when the new data is loading and it
is replaces in place, which means users shouldn't see the data they are
reading jump around.